% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_project.R
\name{init_project}
\alias{init_project}
\title{\code{init_project} creates a series of standard folders for a new project}
\usage{
init_project(use = "science")
}
\value{
a bunch of folders
}
\description{
\code{init_project} creates a series of standard folders for a new project
}
\examples{
init_project()
}

