% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leads_and_lags.R
\name{leads_and_lags}
\alias{leads_and_lags}
\title{leads_and_lags can be used to add in leads or lags of a variables}
\usage{
leads_and_lags(df, variable, foo, x)
}
\arguments{
\item{df}{data frame}

\item{variable}{the variable to lead or lag}

\item{foo}{the function to use (lead or lag)}

\item{x}{the number of leads or lags to include}
}
\value{
a leaded or lagged vector to data frame
}
\description{
leads_and_lags can be used to add in leads or lags of a variables
}
\examples{
edat <- edat \%>\%
group_by(region,fishery,gear) \%>\%
 arrange(year) \%>\%
 leads_and_lags(quo(`average price`),lag,0:5) \%>\%
 leads_and_lags(quo(`average price`),lead,1:5)
}
