% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_functions}
\alias{load_functions}
\title{\code{load_functions} loads locally created functions to the environment}
\usage{
load_functions(func_dir = "functions")
}
\arguments{
\item{func_dir}{the directory where the functions live}
}
\description{
\code{load_functions} loads locally created functions to the environment
}
\examples{
load_functions(func_dir = 'functions')
}
