% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tmaps.R
\name{save_tmaps}
\alias{save_tmaps}
\title{save_tmaps saves all plots in the environment tagged
with the appropriate variable name, such as '_plot'}
\usage{
save_tmaps(plot_dir = ".", extension = ".pdf", plot_tag = "_map",
  fig_width = 6, fig_height = 4)
}
\arguments{
\item{plot_dir}{the directory to save plots}

\item{extension}{the extension to save the plots as}

\item{plot_tag}{the tag to identify what's a plot}

\item{fig_width}{the height of the figure in inches}

\item{fig_height}{the width of the figure in inches}
}
\value{
saved plots
}
\description{
save_tmaps saves all plots in the environment tagged
with the appropriate variable name, such as '_plot'
}
\examples{
save_tmaps(plot_dir = 'results/')
}
